Updating to new Volume IO

+++ bicpl_clapack/dlamch.c
@@ -1,3 +1,5 @@
+#include <stdio.h>
+
 #include "blaswrap.h"
 #include "f2c.h"
 
@@ -592,7 +594,7 @@ w;
 	if (iwarn) {
 	    first = TRUE_;
 	    printf("\n\n WARNING. The value EMIN may be incorrect:- ");
-	    printf("EMIN = %8i\n",lemin);
+	    printf("EMIN = %8i\n",(int)lemin);
 	    printf("If, after inspection, the value EMIN looks acceptable");
             printf("please comment out \n the IF block as marked within the"); 
             printf("code of routine DLAMC2, \n otherwise supply EMIN");
